backend/requirements.txt
fastapi 
uvicorn 
tensorflow 
python-dotenv 
google-generativeai 
numpy 
pandas
scikit-learn